backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
supabase==2.11.0
google-generativeai>=0.8.0
PyMuPDF==1.25.1
pdf2image==1.17.0
Pillow==11.1.0
python-dotenv==1.0.1
python-dateutil==2.9.0
httpx==0.28.1
pydantic==2.10.4
google-cloud-vision>=3.7.0
requests==2.32.3
slowapi==0.1.9
sentry-sdk[fastapi]
gspread>=6.0.0
google-auth>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
